[spring-framework] Apply patches in the Dockerfile
Applying patches is part of the sourcecode retrieval, which
we do in the Dockerfile, so apply the patches as part of the
Docker image creation, too.

While here, compile and install _all_ java files rather then
only those ending on "Fuzzer.java", and use a "real" git
checkout, which is required for more "advanced" git features,
most notoriously "log".